💌 happy yule (winter solstice) - a manifesto ⤵️
during the longest night of the year, sipping some hot cocoa, cozy by the fireplace, i close my eyes and feel so grateful...

remember when i said last month that we are entering a new (20 years) era, with pluto (the planet of death and rebirth and deep transformations) entering aquarius? i can feel it already on my skin, steering up my first house of "self". not everyone has the chance to have this transit in their lifetime - but i will live it intensively during the next 2 decades, and you know i loooove intensity =D.

i am so so so so so excited for my next big thing - won't tell you yet, but it's big, and awesome, and perfect. and it has to do with the "hypergraph" things from urani, and many of the projects i have been working on these last years, together with the best and kindest people i could ask for <3. the universe is always so nice to me, and i live and i have lived the best life i could ask for. i am so grateful for everything. good things happen to good people, karma is real (so if you are not there yet, don't fret, it will come!).

this year was intense, lots of fun adventures and, sure, some painful stuff. being gang-stalked and bullied by freaks on the internet is such a weird concept of our times. but i am a scientist so i framed it as an educational experiment. and it turns out that killing evil is an amazing feeling. all my experimentations and projects has been noteworthy. plus pushing people to be fearless and free is awesome. ah, yeah, and that is also the fact that i faced death and i defeated it - and this only gave me more strength, wisdom, and stamina. 

i never felt this powerful and fearless and judicious in my entire life. i feel like neo, like a "catch-me-if-you-can" ninja, like i am unstoppable in our boundless reality.

i am so excited for 2025, i have some solid plans. many old projects have been retired as they served their purpose (but my community is only getting stronger - the cult has only started =p). besides what i am building (soon, folks!), some new trips, some personal projects. plus the "i feel like i am getting close to meeting my soulmate and stopping being an incel" stuff =p. plus the "i am the hottest and poshest i ever been" stuff. plus the "i am starting a new life in a new country and my apartment is going to be so fuckin awesome and i am throwing the best parties" stuff.

the new year is almost here. even though the world is a little dark atm, the magick is still real if you change your beat for it. so how about you give it a try: open your heart to new feelings, to love*, to hope, to the infinite possibilities of you.

happy holidays, anon
see you on the flip side


* loyal.love is the new project replacing saturnus.tv - you folks know i love gabbing in my free time (: